requirements.txt
streamlit
pandas
numpy
opencv-python-headless
matplotlib
seaborn
requests
python-dotenv
fastapi
uvicorn
torch
torchvision
tensorflow-cpu
pillow
python-multipart
python-dotenv
locust
